Reworked some biome generation to make it look more natural.
