+++ gradle.properties
@@ -5,7 +5,7 @@ org.gradle.jvmargs=-Xmx3G
 org.gradle.daemon=false
 # Stuff for build.gradle.kts
 # info
-mod_version=1.3.3-alpha
+mod_version=1.3.4-alpha
 # https://maven.apache.org/guides/mini/guide-naming-conventions.html
 maven_group=com.crypticcosmos.crypticcosmos
 # File name for jar + '-' + version
