+++ src/main/java/com/crypticcosmos/crypticcosmos/blocks/InfectableBlock.java
@@ -2,41 +2,40 @@ package com.crypticcosmos.crypticcosmos.blocks;
 
 import net.minecraft.block.Block;
 import net.minecraft.block.BlockState;
-import net.minecraft.block.Blocks;
 import net.minecraft.item.BlockItemUseContext;
 import net.minecraft.state.StateContainer;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.server.ServerWorld;
 
+import javax.annotation.Nonnull;
 import java.util.Random;
 
-@SuppressWarnings("NullableProblems")
-public class MoonPlanks extends Block implements Infectable {
-    public MoonPlanks() {
-        super(Properties.copy(Blocks.BIRCH_PLANKS));
+public class InfectableBlock extends Block implements Infectable {
+    public InfectableBlock(Properties properties) {
+        super(properties);
 
         this.registerDefaultState(this.defaultBlockState().setValue(INFECTION_LEVEL, 0));
     }
 
     @SuppressWarnings("deprecation")
     @Override
-    public void randomTick(BlockState state, ServerWorld world, BlockPos pos, Random rand) {
+    public void randomTick(@Nonnull BlockState state, @Nonnull ServerWorld world, @Nonnull BlockPos pos, @Nonnull Random rand) {
         this.infect(world, pos);
     }
 
     @Override
-    protected void createBlockStateDefinition(StateContainer.Builder<Block, BlockState> builder) {
+    protected void createBlockStateDefinition(@Nonnull StateContainer.Builder<Block, BlockState> builder) {
         super.createBlockStateDefinition(builder);
         builder.add(INFECTION_LEVEL);
     }
 
     @Override
-    public BlockState getStateForPlacement(BlockItemUseContext context) {
+    public BlockState getStateForPlacement(@Nonnull BlockItemUseContext context) {
         return this.defaultBlockState().setValue(INFECTION_LEVEL, 0);
     }
 
     @Override
-    public boolean isRandomlyTicking(BlockState state) {
+    public boolean isRandomlyTicking(@Nonnull BlockState state) {
         return true;
     }
 }
