+++ src/main/java/com/crypticcosmos/crypticcosmos/registries/BlockRegistries.java
@@ -2,13 +2,18 @@ package com.crypticcosmos.crypticcosmos.registries;
 
 import com.crypticcosmos.crypticcosmos.CrypticCosmos;
 import com.crypticcosmos.crypticcosmos.blocks.*;
+import com.crypticcosmos.crypticcosmos.world.feature.MondroveTree;
 import net.minecraft.block.AbstractBlock.Properties;
 import net.minecraft.block.*;
 import net.minecraft.block.material.Material;
+import net.minecraft.util.math.BlockPos;
+import net.minecraft.world.IBlockReader;
 import net.minecraftforge.common.ToolType;
 import net.minecraftforge.fml.RegistryObject;
 import net.minecraftforge.registries.DeferredRegister;
 import net.minecraftforge.registries.ForgeRegistries;
+
+import javax.annotation.Nonnull;
 
 public class BlockRegistries {
     public static final DeferredRegister<Block> BLOCKS = DeferredRegister.create(ForgeRegistries.BLOCKS, CrypticCosmos.MOD_ID);
@@ -84,11 +89,19 @@ public class BlockRegistries {
     );
 
     //Moon Wood
-    public static final RegistryObject<Block> MONDROVE_LOG = BLOCKS.register("mondrove_log", MoonLog::new);
+    public static final RegistryObject<Block> MONDROVE_LOG = BLOCKS.register("mondrove_log", MondroveLog::new);
 
-    public static final RegistryObject<Block> MONDROVE_PLANKS = BLOCKS.register("mondrove_planks", MoonPlanks::new);
+    public static final RegistryObject<Block> MONDROVE_PLANKS = BLOCKS.register("mondrove_planks", () ->
+            new InfectableBlock(Properties.copy(Blocks.BIRCH_PLANKS))
+    );
 
-    public static final RegistryObject<Block> MONDROVE_SAPLING = BLOCKS.register("mondrove_sapling", MoonSapling::new);
+    public static final RegistryObject<Block> MONDROVE_SAPLING = BLOCKS.register("mondrove_sapling", () ->
+            new SaplingBlock(new MondroveTree(), Properties.copy(Blocks.BIRCH_SAPLING)) {
+                @Override
+                public boolean mayPlaceOn(@Nonnull BlockState state, @Nonnull IBlockReader worldIn, @Nonnull BlockPos pos) {
+                    return state.is(TagRegistries.LUNARA_PLANTABLE_BLOCKS);
+                }
+            });
 
     public static final RegistryObject<Block> MONDROVE_LEAVES = BLOCKS.register("mondrove_leaves", () ->
             new LeavesBlock(Properties.copy(Blocks.OAK_LEAVES))
