+++ src/main/java/com/crypticcosmos/crypticcosmos/world/biomes/BiomeMaker.java
@@ -4,24 +4,24 @@ import com.crypticcosmos.crypticcosmos.registries.BlockRegistries;
 import com.crypticcosmos.crypticcosmos.registries.EntityTypeRegistries;
 import com.crypticcosmos.crypticcosmos.registries.FeatureRegistries;
 import com.crypticcosmos.crypticcosmos.registries.SoundEventRegistries;
-import net.minecraft.client.audio.BackgroundMusicTracks;
 import net.minecraft.entity.EntityClassification;
 import net.minecraft.entity.EntityType;
-import net.minecraft.util.SoundEvent;
-import net.minecraft.util.math.MathHelper;
 import net.minecraft.world.biome.Biome;
-import net.minecraft.world.biome.Biome.Builder;
 import net.minecraft.world.biome.Biome.Category;
 import net.minecraft.world.biome.Biome.RainType;
 import net.minecraft.world.biome.BiomeAmbience;
 import net.minecraft.world.biome.BiomeGenerationSettings;
 import net.minecraft.world.biome.MobSpawnInfo;
 import net.minecraft.world.gen.GenerationStage;
-import net.minecraft.world.gen.surfacebuilders.ISurfaceBuilderConfig;
+import net.minecraft.world.gen.feature.Features;
+import net.minecraft.world.gen.placement.AtSurfaceWithExtraConfig;
+import net.minecraft.world.gen.placement.Placement;
 import net.minecraft.world.gen.surfacebuilders.SurfaceBuilder;
 import net.minecraft.world.gen.surfacebuilders.SurfaceBuilderConfig;
 
 import javax.annotation.Nonnull;
+
+import static com.crypticcosmos.crypticcosmos.world.biomes.BiomeHelper.*;
 
 @SuppressWarnings("SameParameterValue")
 public class BiomeMaker {
@@ -37,15 +37,7 @@ public class BiomeMaker {
             BlockRegistries.UMBRAL_DUNE.get().defaultBlockState()
     );
 
-    @SuppressWarnings("unused")
-    public static final int DEFAULT_WATER_COLOR = 0x3f76e4;
-    @SuppressWarnings("unused")
-    public static final int DEFAULT_WATER_FOG_COLOR = 0x50533;
-    public static final int DEFAULT_GRASS_COLOR = 0x91bd59;
-    public static final int DEFAULT_FOLIAGE_COLOR = 0x77ab2f;
-    public static final int DEFAULT_SKY_FOG_COLOR = 12638463;
-
-    public static Biome makeAcerbicIsles() {
+    public static Biome acerbicIsles() {
         final BiomeGenerationSettings.Builder genSettings = genSettings(SurfaceBuilder.DEFAULT, LUNARA_SURFACE_BUILDER_CONFIG);
 
         final MobSpawnInfo.Builder spawnSettings = spawnSettings();
@@ -78,14 +70,22 @@ public class BiomeMaker {
         );
     }
 
-    public static Biome makeMondroveGroves() {
+    public static Biome mondroveGroves() {
         final BiomeGenerationSettings.Builder genSettings = genSettings(SurfaceBuilder.DEFAULT, LUNARA_SURFACE_BUILDER_CONFIG);
 
         // Add mondrove fungus generation.
-        genSettings.addFeature(GenerationStage.Decoration.VEGETAL_DECORATION, FeatureRegistries.MONDROVE_FUNGUS);
+        genSettings.addFeature(GenerationStage.Decoration.VEGETAL_DECORATION,
+                FeatureRegistries.MONDROVE_FUNGUS
+                        .decorated(Features.Placements.ADD_32)
+                        .decorated(Features.Placements.HEIGHTMAP_SQUARE)
+                        .count(2));
 
         // Add mondrove tree generation.
-        genSettings.addFeature(GenerationStage.Decoration.VEGETAL_DECORATION, FeatureRegistries.MONDROVE_TREE);
+        genSettings.addFeature(GenerationStage.Decoration.VEGETAL_DECORATION,
+                FeatureRegistries.MONDROVE_TREE
+                        .decorated(Features.Placements.HEIGHTMAP_SQUARE)
+                        .decorated(Placement.COUNT_EXTRA
+                                .configured(new AtSurfaceWithExtraConfig(7, 0.1F, 1))));
 
         final MobSpawnInfo.Builder spawnSettings = spawnSettings();
 
@@ -105,7 +105,7 @@ public class BiomeMaker {
                 RainType.RAIN,
                 Category.FOREST,
                 0.125f,
-                1,
+                0.07f,
                 0f,
                 0.0001f,
                 effects,
@@ -115,7 +115,7 @@ public class BiomeMaker {
     }
 
     @Nonnull
-    public static Biome makeLunaraPlains() {
+    public static Biome lunaraPlains() {
         final BiomeGenerationSettings.Builder genSettings = genSettings(SurfaceBuilder.DEFAULT, LUNARA_SURFACE_BUILDER_CONFIG);
 
         final MobSpawnInfo.Builder spawnSettings = spawnSettings();
@@ -138,7 +138,7 @@ public class BiomeMaker {
                 RainType.RAIN,
                 Category.PLAINS,
                 0.125f,
-                1,
+                0.1f,
                 0f,
                 0.0001f,
                 effects,
@@ -147,7 +147,7 @@ public class BiomeMaker {
         );
     }
 
-    public static Biome makeUmbralDunes() {
+    public static Biome umbralDunes() {
         final BiomeGenerationSettings.Builder genSettings = genSettings(SurfaceBuilder.DEFAULT, ABYSS_SURFACE_BUILDER_CONFIG);
 
         final BiomeAmbience.Builder effects = effects(0x412,
@@ -169,85 +169,5 @@ public class BiomeMaker {
                 genSettings,
                 MobSpawnInfo.EMPTY
         );
-    }
-
-    /**
-     * Base biome function Sky color is not generated
-     */
-    private static Biome biome(
-            RainType precipitation,
-            Category category,
-            float depth,
-            float scale,
-            float temperature,
-            float downfall,
-            BiomeAmbience.Builder effects,
-            BiomeGenerationSettings.Builder genSettings,
-            MobSpawnInfo spawnSettings
-    ) {
-        return new Builder()
-                .precipitation(precipitation)
-                .biomeCategory(category)
-                .depth(depth)
-                .scale(scale)
-                .temperature(temperature)
-                .downfall(downfall)
-                .specialEffects(effects.build())
-                .generationSettings(genSettings.build())
-                .mobSpawnSettings(spawnSettings)
-                .build();
-    }
-
-    /**
-     * Shortcut function and enforces surface builder
-     */
-    private static <C extends ISurfaceBuilderConfig> BiomeGenerationSettings.Builder genSettings(SurfaceBuilder<C> surfaceBuilder, C config) {
-        return new BiomeGenerationSettings.Builder().surfaceBuilder(surfaceBuilder.configured(config));
-    }
-
-    /**
-     * Shortcut function
-     */
-    private static MobSpawnInfo.Builder spawnSettings() {
-        return new MobSpawnInfo.Builder();
-    }
-
-    /**
-     * Shortcut function
-     */
-    private static void addSpawn(
-            MobSpawnInfo.Builder spawnSettings,
-            EntityClassification classification,
-            EntityType<?> entityType,
-            int weight,
-            int min,
-            int max) {
-        spawnSettings.addSpawn(classification,
-                new MobSpawnInfo.Spawners(entityType, weight, min, max));
-    }
-
-    /**
-     * Biome ambience add default parameters and enforced the required ones. Should prevent slip ups on my part :)
-     */
-    private static BiomeAmbience.Builder effects(int waterColor,
-                                                 int waterFogColor,
-                                                 int grassColor,
-                                                 int foliageColor,
-                                                 float temperature,
-                                                 int skyFogColor,
-                                                 SoundEvent music) {
-        return new BiomeAmbience.Builder()
-                .waterColor(waterColor)
-                .waterFogColor(waterFogColor)
-                .grassColorOverride(grassColor)
-                .foliageColorOverride(foliageColor)
-                .skyColor(getSkyForTemp(temperature))
-                .fogColor(skyFogColor)
-                .backgroundMusic(BackgroundMusicTracks.createGameMusic(music));
-    }
-
-    private static int getSkyForTemp(float temperature) {
-        final float a = MathHelper.clamp(temperature / 3.0f, -1.0f, 1.0f);
-        return MathHelper.hsvToRgb(0.62222224f - a * 0.05f, 0.5f + a * 0.1f, 1.0f);
     }
 }
