+++ src/main/resources/META-INF/accesstransformer.cfg
@@ -2,3 +2,4 @@ public net.minecraft.data.loot.BlockLootTables field_218579_g # DEFAULT_SAPLING_
 public net.minecraft.block.Blocks func_235430_a_(Lnet/minecraft/block/material/MaterialColor;Lnet/minecraft/block/material/MaterialColor;)Lnet/minecraft/block/RotatedPillarBlock; # createLogBlock
 public net.minecraft.data.loot.BlockLootTables func_218526_b(Lnet/minecraft/block/Block;Lnet/minecraft/block/Block;[F)Lnet/minecraft/loot/LootTable$Builder; # droppingWithChancesSticksAndApples
 public net.minecraft.item.DebugStickItem func_195960_a(Lnet/minecraft/block/BlockState;Lnet/minecraft/state/Property;Z)Lnet/minecraft/block/BlockState; # cycleProperty
+public net.minecraft.item.SpawnEggItem field_195987_b # EGGS
