+++ src/main/resources/META-INF/mods.toml
@@ -7,7 +7,7 @@ issueTrackerURL = "https://github.com/Team-Cryptic-Cosmos/Cryptic-Cosmos/issues"
 [[mods]]
 modId = "crypticcosmos"
 displayName = "Cryptic Cosmos"
-version = "1.3.2-alpha"
+version = "1.3.4-alpha"
 description = """
 This mod takes inspiration from the writer H.P Lovecraft.
 Here, we make all sorts of new mobs and dimensions where these monsters come from.
