+++ src/main/resources/data/crypticcosmos/worldgen/noise_settings/abyss_settings.json
@@ -1,7 +1,7 @@
 {
   "bedrock_roof_position": -10,
   "bedrock_floor_position": 0,
-  "sea_level": 63,
+  "sea_level": 41,
   "disable_mob_generation": false,
   "structures": {
     "structures": {}
